requirements.txt
python-dotenv
requests
httpx
moviepy==1.0.3
pyttsx3
srt
numpy
rich
Pillow>=10